# contracts/openapi/intelliQuiz.v1.yaml
openapi: 3.1.0
info:
  title: IntelliQuiz API
  version: 1.0.0
  description: |
    Central API for IntelliQuiz. Clients: Admin Web, Student Desktop.
    Android pairs with Desktop locally and does not call most of these endpoints.
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
tags:
  - name: health
  - name: auth
  - name: exams
  - name: sessions
  - name: sync
  - name: reports
paths:
  /health:
    get:
      tags: [health]
      summary: Liveness probe
      operationId: getHealth
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /auth/login:
    post:
      tags: [auth]
      summary: Exchange credentials for access token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid credentials

  /exams:
    get:
      tags: [exams]
      summary: List exams for current tenant
      operationId: listExams
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Exam list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamSummary"
    post:
      tags: [exams]
      summary: Create exam
      operationId: createExam
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamSummary"

  /exams/{exam_id}:
    get:
      tags: [exams]
      summary: Get exam detail
      operationId: getExam
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ExamId"
      responses:
        "200":
          description: Exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamDetail"

  /sessions:
    post:
      tags: [sessions]
      summary: Start a student exam session
      operationId: createSession
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionCreate"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"

  /sessions/{session_id}/heartbeat:
    post:
      tags: [sessions]
      summary: Desktop heartbeat
      operationId: sessionHeartbeat
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeartbeatRequest"
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  /sync/events:
    post:
      tags: [sync]
      summary: Idempotent integrity event batch upload
      operationId: syncEvents
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventBatch"
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"

  /reports/sessions/{session_id}:
    get:
      tags: [reports]
      summary: Integrity report for a session
      operationId: getSessionReport
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrityReport"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    ExamId:
      name: exam_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SessionId:
      name: session_id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  schemas:
    HealthResponse:
      type: object
      required: [status, service, version]
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
          example: intelliquiz-api
        version:
          type: string
          example: 1.0.0

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    TokenResponse:
      type: object
      required: [access_token, token_type, role]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer
        role:
          type: string
          enum: [admin, instructor, student]
        expires_in:
          type: integer
          example: 3600

    ExamSummary:
      type: object
      required: [id, title, status, duration_minutes]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
        duration_minutes:
          type: integer
        proctoring_profile_id:
          type: string
          format: uuid
          nullable: true

    ExamCreate:
      type: object
      required: [title, duration_minutes]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        duration_minutes:
          type: integer
          minimum: 1
          maximum: 600
        proctoring_profile_id:
          type: string
          format: uuid
          nullable: true
        instructions:
          type: string

    ExamDetail:
      allOf:
        - $ref: "#/components/schemas/ExamSummary"
        - type: object
          properties:
            instructions:
              type: string
            question_count:
              type: integer

    SessionCreate:
      type: object
      required: [exam_id, device_fingerprint]
      properties:
        exam_id:
          type: string
          format: uuid
        device_fingerprint:
          type: string

    SessionResponse:
      type: object
      required: [id, exam_id, status, started_at, pairing_token]
      properties:
        id:
          type: string
          format: uuid
        exam_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, active, submitted, synced]
        started_at:
          type: string
          format: date-time
        pairing_token:
          type: string
          description: Short-lived secret embedded in Desktop QR for Android

    HeartbeatRequest:
      type: object
      required: [client_ts, risk_score]
      properties:
        client_ts:
          type: string
          format: date-time
        risk_score:
          type: number
          minimum: 0
          maximum: 1
        android_paired:
          type: boolean

    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean

    EventBatch:
      type: object
      required: [session_id, events]
      properties:
        session_id:
          type: string
          format: uuid
        events:
          type: array
          minItems: 1
          maxItems: 500
          items:
            $ref: "#/components/schemas/IntegrityEvent"

    IntegrityEvent:
      type: object
      required: [event_id, type, ts, severity, payload]
      properties:
        event_id:
          type: string
          format: uuid
        type:
          type: string
          enum:
            - gesture_risk
            - gaze_away
            - multi_face
            - no_face
            - app_violation
            - android_env_anomaly
            - heartbeat
            - submit
        ts:
          type: string
          format: date-time
        severity:
          type: number
          minimum: 0
          maximum: 1
        payload:
          type: object
          additionalProperties: true

    SyncResult:
      type: object
      required: [accepted, duplicates, rejected]
      properties:
        accepted:
          type: integer
        duplicates:
          type: integer
        rejected:
          type: integer

    IntegrityReport:
      type: object
      required: [session_id, cheating_probability, flags, event_count]
      properties:
        session_id:
          type: string
          format: uuid
        cheating_probability:
          type: number
          minimum: 0
          maximum: 1
        flags:
          type: array
          items:
            type: string
        event_count:
          type: integer
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/IntegrityEvent"